Add ResourceLink as a page part
