+++ app/views/spina/admin/resource_select_options/index.html.erb
@@ -1,0 +1,13 @@
+<%= turbo_frame_tag "resource_select_options_#{params[:object_id]}" do %>
+  <% if params[:search].blank? %>
+    <button type="button" data-action="select#clear reveal#hide" class="block hover:bg-gray-100 w-full text-left px-5 py-2 font-medium leading-5 text-gray-700 hover:text-gray-900 text-sm border-t border-gray-150">
+      –
+    </button>
+  <% end %>
+
+  <% @resources.each do |resource| %>
+    <button type="button" data-action="select#select reveal#hide" data-title="<%= resource.label %>" data-id="<%= resource.id %>" class="block hover:bg-gray-100 w-full text-left px-5 py-2 font-medium leading-5 text-gray-700 hover:text-gray-900 text-sm border-t border-gray-150">
+      <%= resource.label %>
+    </button>
+  <% end %>
+<% end %>
