+++ app/views/spina/admin/resource_select_options/show.html.erb
@@ -1,0 +1,3 @@
+<%= turbo_frame_tag :resource_title do %>
+  <%= @resource.label %>
+<% end %>
