+++ docs/v2/rendering_content/10_resource_link.md
@@ -1,0 +1,16 @@
+# ResourceLink
+
+An ResourceLink is a part where you can link to or reference a resource. This is useful for displaying a list of pages contained within a resource.
+
+## Theme configuration
+
+```
+config.parts = [
+  # ...
+  {
+    name: "content_group",
+    title: "Content Group",
+    part_type: "Spina::Parts::ResourceLink"
+  }
+]
+```
